Split out terms_test
Split legislative period tests into their own test file
